pyproject.toml
[project]
name = "RAG_tron"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Collin Barber", email = "collinjbarber@gmail.com"},
]
dependencies = [
    "openai>=1.51.0",
    "requests>=2.32.3",
    "markdownify>=0.13.1",
    "bs4>=0.0.2",
    "langchain>=0.3.2",
    "pinecone-client>=5.0.1",
    "discord-py>=2.4.0",
    "pydantic-settings>=2.5.2",
    "langchain-community>=0.3.1",
    "langchain-openai>=0.2.2",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
start = "python -m src.main"
embed = "python -m src.scrape"